refactor: TryCommand 패키지 이동, 조회 로직 변경
